requirements.txt
python-telegram-bot==20.*
apscheduler==3.10.*
openai==1.*
python-dotenv
